requirements.txt
fastapi==0.115.6
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.0
uvicorn==0.34.0
python-multipart==0.0.18
pydantic==2.10.3
google-generativeai==0.8.3
